De-dupe on broadcast_address instead of hostname
With docker deployments it is common for the nsqd dockers to have the same hostname.

De-duping on the broadcast address rather than hostname (which is immutable with nsqd options) seems to make more sense. This solves a bug in our prod system where only one nsqd was being listened to at a time due to duplicate hostnames.